TIKR/prod/ml/deploy_cloud_function/requirements.txt
functions-framework==3.4.0
tensorflow==2.13.0
keras
yfinance
google-cloud-storage
joblib
